replaced old DspTester with VerboseDspTester (renamed as DspTester); sbt test passes BUT needs major cleanup
